requirements.txt
py-cord
yaml2object~=1.0.13
asyncclick==8.0.3.1
setuptools==60.5.0
PyYAML==6.0
anyio==3.5.0